+++ CHANGES.rst
@@ -12,7 +12,7 @@ in progress
 - Activate full ``umoci``-unpacked images
 - Add test case for ``postroj pkgprobe``
 - Improve image subsystem to acquire arbitrary Docker filesystem images
-- Add operating system support for SLES, RHEL
+- Add operating system support for SLES, RHEL, CentOS 9
 
 
 2022-05-08 0.1.0